구조 정리 디렉터리 정리, Container component 작성
