+++ src/components/App/App.test.jsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import { useSelector } from 'react-redux';
 
-import App from './App';
+import App from '.';
 
 jest.mock('react-redux');
