+++ src/components/App/index.jsx
@@ -1,0 +1,12 @@
+import InputContainer from '../InputContainer';
+import ListContainer from '../ListContainer';
+
+export default function App() {
+  return (
+    <div>
+      <h1>To-do</h1>
+      <InputContainer />
+      <ListContainer />
+    </div>
+  );
+}
