+++ src/index.jsx
@@ -1,7 +1,7 @@
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
-import App from './App';
+import App from './components/App';
 import store from './store/store';
 
 ReactDOM.render(
